+++ ios/samples/gltf-viewer/gltf-viewer/FILModelView.mm
@@ -47,7 +47,7 @@
 
 using namespace filament;
 using namespace utils;
-using namespace gltfio;
+using namespace filament::gltfio;
 using namespace camutils;
 
 const double kNearPlane = 0.05;   // 5 cm
@@ -126,7 +126,7 @@ const float kSensitivity = 100.0f;
 
     _swapChain = _engine->createSwapChain((__bridge void*)self.layer);
 
-    _materialProvider = createUbershaderLoader(_engine);
+    _materialProvider = createUbershaderProvider(_engine);
     EntityManager& em = EntityManager::get();
     NameComponentManager* ncm = new NameComponentManager(em);
     _assetLoader = AssetLoader::create({_engine, _materialProvider, ncm, &em});