+++ samples/gltf_viewer.cpp
@@ -63,7 +63,7 @@ using namespace filament;
 using namespace filament::math;
 using namespace filament::viewer;
 
-using namespace gltfio;
+using namespace filament::gltfio;
 using namespace utils;
 
 enum MaterialSource {
@@ -495,7 +495,7 @@ int main(int argc, char** argv) {
         }
 
         app.materials = (app.materialSource == GENERATE_SHADERS) ?
-                createMaterialGenerator(engine) : createUbershaderLoader(engine);
+                createJitShaderProvider(engine) : createUbershaderProvider(engine);
         app.assetLoader = AssetLoader::create({engine, app.materials, app.names });
         app.mainCamera = &view->getCamera();
         if (filename.isEmpty()) {